Support exclusion of specific tiles from Image monitoring range tests
Feature provides a way to exclude particular tiles from image monitoring.
This is beneficial if 'all' or many tiles are in ROI, but some set ot tiles
should be not included. Defining the full list of desired monitoring tiles
would be time-consuming. By introducing new cell called 'Tiles to be excluded'
in Image monitoring settings, one can simply list those specific tiles that
will be not checked for Mean/SD range and Slice-by-salice threshold tests.
Mind that tiles to be excluded must be a sub-set of either 'all' or list of
tile keys in 'Tiles to be monitored' cell.