AFSS: Background mode implemented  (no AFSS corrections applied)
Background mode will not apply any of the optimal values at the end of AFSS series.
It can serve as a method to track the Focus/Stigmation evolution manually (using
log messages and/or stats files so that user can decide when and what changes to
Focus/Stig values will be applied. No impact on any other functionality.
Can be activated deactivted from AFSS Dlg Win.